# openapi/openapi.yaml
openapi: 3.1.0
info:
  description: "Online task planning system — manage tasks, work slots, dependencies\
    \ and generate optimized plans."
  title: ChronoScope API
  version: 1.1.0
servers:
- description: Production
  url: https://chronoscope.ni0.team/api
security:
- oauth2: []
...
paths:
  /v1/identity:
    get:
      description: "Retrieve information about the authenticated identity, including\
        \ all linked accounts and the organizations for which the identity has admin\
        \ privileges, as resolved server-side."
      operationId: getIdentity
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdentityResponse"
          description: Identity retrieved successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
      summary: Get current identity
      tags:
      - Identity
...
  /v1/organizations/{organizationId}/invitations:
    get:
      description: Returns all pending invitations for the organization. The identity
        must be an admin of the organization.
      operationId: getInvitations
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrganizationInvitationsResponse"
          description: Invitations retrieved successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not an admin of the organization
      summary: List organization invitations
      tags:
      - Organization
  /v1/organizations/{organizationId}/invitations/{invitationId}:
    delete:
      description: Deletes a pending invitation. The identity must be an admin of
        the organization.
      operationId: deleteInvitation
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        schema:
          type: string
      - description: Invitation ID
        in: path
        name: invitationId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Invitation deleted
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not an admin of the organization
      summary: Delete organization invitation
      tags:
      - Organization
    put:
      description: Resends an existing invitation email. The identity must be an admin
        of the organization.
      operationId: resendInvitation
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        schema:
          type: string
      - description: Invitation ID
        in: path
        name: invitationId
        required: true
        schema:
          type: string
      responses:
        "202":
          description: Invitation resent
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not an admin of the organization
      summary: Resend organization invitation
      tags:
      - Organization
  /v1/organizations/{organizationId}/members:
    get:
      description: Retrieves members of a specific organization.
      operationId: getOrganizationMembers
      parameters:
      - in: path
        name: organizationId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrganizationMembersResponse"
          description: Members retrieved successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not an admin of the organization
      summary: Get Organization members
      tags:
      - Organization
    post:
      description: Sends an invitation email to the given address. The identity must
        be an admin of the organization.
      operationId: inviteUser
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserRequest"
        required: true
      responses:
        "202":
          description: Invitation sent
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation error
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not an admin of the organization
      summary: Invite user to organization
      tags:
      - Organization
  /v1/organizations/{organizationId}/members/{id}:
    delete:
      description: Removes a member from the organization. The identity must be an
        admin of the organization.
      operationId: removeMember
      parameters:
      - description: Organization ID
        in: path
        name: organizationId
        required: true
        schema:
          type: string
      - description: User ID to remove
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Member removed successfully
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not an admin of the organization
      summary: Remove organization member
      tags:
      - Organization
  /v1/plan:
    post:
      description: Run the planning algorithm for the given account. Returns planned
        task scopes.
      operationId: plan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScopeResponse"
          description: Plan generated successfully
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation error
        "401":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid token
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Not enough work slots to accommodate all tasks
      summary: Generate plan
      tags:
      - Planning
...
components:
  schemas:
...
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        identityId:
          type: integer
          format: int64
      required:
      - id
      - identityId
...
    IdentityResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        adminOrganizations:
          type: array
          description: "Organization IDs for which the authenticated identity has\
            \ admin privileges, resolved server-side via Keycloak Admin Client lookups."
          items:
            type: string
          uniqueItems: true
        id:
          type: integer
          format: int64
        organizations:
          type: array
          description: Organizations to which the identity belongs.
          items:
            $ref: "#/components/schemas/Organization"
          uniqueItems: true
      required:
      - accounts
      - adminOrganizations
      - id
      - organizations
    Invitation:
      type: object
      properties:
        email:
          type: string
        expired:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        firstName:
          type: string
        id:
          type: string
        inviteLink:
          type: string
        lastName:
          type: string
      required:
      - email
      - expired
      - expiresAt
      - firstName
      - id
      - inviteLink
      - lastName
    InviteUserRequest:
      type: object
      properties:
        targetMail:
          type: string
          format: email
          minLength: 1
      required:
      - targetMail
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    OrganizationInvitationsResponse:
      type: object
      properties:
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/Invitation"
      required:
      - invitations
    OrganizationMember:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        id:
          type: string
        lastName:
          type: string
        userName:
          type: string
      required:
      - email
      - firstName
      - id
      - lastName
      - userName
    OrganizationMembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMember"
      required:
      - members
    PlanRequest:
      type: object
      properties:
        organizationId:
          type: string
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
    ScopeResponse:
      type: object
      properties:
        dynamicTaskId:
          type: integer
          format: int64
        endAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        startAt:
          type: string
          format: date-time
      required:
      - dynamicTaskId
      - endAt
      - id
      - startAt
...
  securitySchemes:
    oauth2:
      description: Keycloak OAuth2 (Authorization Code flow)
      flows:
        authorizationCode:
          authorizationUrl: https://auth.ni0.team/realms/ni0/protocol/openid-connect/auth
          scopes:
            read: Read access
            write: Write access
          tokenUrl: https://auth.ni0.team/realms/ni0/protocol/openid-connect/token
      type: oauth2