# src/main/java/de/ni0/chronoscope/controller/dto/request/InviteUserRequest.java
package de.ni0.chronoscope.controller.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record InviteUserRequest(
        @NotBlank @Email String targetMail
) {
}